Map Rust u8/u16 to Go uint8/uint16 instead of signed int
The Go generator folded u8 and u16 into the signed `int` arm, while u32 maps
to `uint32` and u64 to `uint64`. This silently drops the unsignedness for the
two smallest unsigned types, which is internally inconsistent and lets a Go
service place a negative or out-of-range value into a field that is unsigned on
the Rust side — serde then rejects it on deserialization.

The repo's own `use_correct_integer_types` fixture already establishes the
intended contract: for the same u8/u16 inputs, Swift emits UInt8/UInt16 and
Kotlin emits UByte/UShort. Map u8 -> uint8 and u16 -> uint16 so Go matches.

Regenerated the Go snapshots; the only changes are int -> uint8/uint16 for u8/u16
fields (i32 fields stay `int`). Full typeshare-core suite passes.

+++ core/data/tests/use_correct_integer_types/output.go
@@ -7,7 +7,7 @@ type Foo struct {
 	A int `json:"a"`
 	B int `json:"b"`
 	C int `json:"c"`
-	E int `json:"e"`
-	F int `json:"f"`
+	E uint8 `json:"e"`
+	F uint16 `json:"f"`
 	G uint32 `json:"g"`
 }
